Fix findings: module structure, footnotes, and validation
- Split where_clause.rs into where_clause/mod.rs and where_clause/metavariable.rs
  to satisfy clippy::self-named-module-files lint
- Add field validation to parse_metavariable_clause with proper error handling
- Add semantic validation in parse_metavariable_pattern_clause using validate_formula_semantics
- Update build_match_formula_from_mapping to handle decoration keys where/as/fix
- Fix weaver-design.md footnote numbering: assign unique sequential numbers
